More fixes (no tests yet)
