+++ src/bazel.rs
@@ -58,12 +58,18 @@ impl BazelRenderer {
           "templates/partials/remote_rust_library.template",
           include_str!("templates/partials/remote_rust_library.template"),
         ),
-        ("templates/workspace.BUILD.template", include_str!("templates/workspace.BUILD.template")),
+        (
+          "templates/workspace.BUILD.template",
+          include_str!("templates/workspace.BUILD.template"),
+        ),
         (
           "templates/remote_workspace.BUILD.template",
           include_str!("templates/remote_workspace.BUILD.template"),
         ),
-        ("templates/crate.BUILD.template", include_str!("templates/crate.BUILD.template")),
+        (
+          "templates/crate.BUILD.template",
+          include_str!("templates/crate.BUILD.template"),
+        ),
         (
           "templates/remote_crates.bzl.template",
           include_str!("templates/remote_crates.bzl.template"),
@@ -88,7 +94,9 @@ impl BazelRenderer {
     let mut context = Context::new();
     context.add("workspace", &workspace_context);
     context.add("crate", &package);
-    self.internal_renderer.render("templates/crate.BUILD.template", &context)
+    self
+      .internal_renderer
+      .render("templates/crate.BUILD.template", &context)
   }
 
   pub fn render_aliases(
@@ -99,7 +107,9 @@ impl BazelRenderer {
     let mut context = Context::new();
     context.add("workspace", &workspace_context);
     context.add("crates", &all_packages);
-    self.internal_renderer.render("templates/workspace.BUILD.template", &context)
+    self
+      .internal_renderer
+      .render("templates/workspace.BUILD.template", &context)
   }
 
   pub fn render_remote_crate(
@@ -110,7 +120,9 @@ impl BazelRenderer {
     let mut context = Context::new();
     context.add("workspace", &workspace_context);
     context.add("crate", &package);
-    self.internal_renderer.render("templates/remote_crate.BUILD.template", &context)
+    self
+      .internal_renderer
+      .render("templates/remote_crate.BUILD.template", &context)
   }
 
   pub fn render_remote_aliases(
@@ -121,7 +133,9 @@ impl BazelRenderer {
     let mut context = Context::new();
     context.add("workspace", &workspace_context);
     context.add("crates", &all_packages);
-    self.internal_renderer.render("templates/remote_workspace.BUILD.template", &context)
+    self
+      .internal_renderer
+      .render("templates/remote_workspace.BUILD.template", &context)
   }
 
   pub fn render_bzl_fetch(
@@ -132,7 +146,9 @@ impl BazelRenderer {
     let mut context = Context::new();
     context.add("workspace", &workspace_context);
     context.add("crates", &all_packages);
-    self.internal_renderer.render("templates/remote_crates.bzl.template", &context)
+    self
+      .internal_renderer
+      .render("templates/remote_crates.bzl.template", &context)
   }
 }
 
@@ -143,8 +159,7 @@ impl BuildRenderer for BazelRenderer {
     planned_build: &PlannedBuild,
   ) -> CargoResult<Vec<FileOutputs>> {
     let &RenderDetails {
-      ref path_prefix,
-      ..
+      ref path_prefix, ..
     } = render_details;
     let &PlannedBuild {
       ref workspace_context,
@@ -185,8 +200,7 @@ impl BuildRenderer for BazelRenderer {
     planned_build: &PlannedBuild,
   ) -> CargoResult<Vec<FileOutputs>> {
     let &RenderDetails {
-      ref path_prefix,
-      ..
+      ref path_prefix, ..
     } = render_details;
     let &PlannedBuild {
       ref workspace_context,
@@ -202,7 +216,11 @@ impl BuildRenderer for BazelRenderer {
     });
 
     for package in crate_contexts {
-      let build_file_path = format!("remote/{}-{}.BUILD", &package.pkg_name, &package.pkg_version);
+      let build_file_path = format!(
+        "remote/{}-{}.BUILD",
+        &package.pkg_name,
+        &package.pkg_version
+      );
       let rendered_crate_build_file = try!(
         self
           .render_remote_crate(&workspace_context, &package)
@@ -344,15 +362,24 @@ mod tests {
   #[test]
   fn all_plans_contain_root_build_file() {
     let file_outputs = render_crates_for_test(Vec::new());
-    let file_names = file_outputs.iter().map(|output| output.path.as_ref()).collect::<Vec<&str>>();
+    let file_names = file_outputs
+      .iter()
+      .map(|output| output.path.as_ref())
+      .collect::<Vec<&str>>();
 
-    assert_that!(&file_names, contains(vec!["./some_render_prefix/BUILD"]).exactly());
+    assert_that!(
+      &file_names,
+      contains(vec!["./some_render_prefix/BUILD"]).exactly()
+    );
   }
 
   #[test]
   fn crates_generate_build_files() {
     let file_outputs = render_crates_for_test(vec![dummy_library_crate()]);
-    let file_names = file_outputs.iter().map(|output| output.path.as_ref()).collect::<Vec<&str>>();
+    let file_names = file_outputs
+      .iter()
+      .map(|output| output.path.as_ref())
+      .collect::<Vec<&str>>();
 
     assert_that!(
       &file_names,
