+++ src/main.rs
@@ -41,10 +41,10 @@ use cargo::CargoError;
 use cargo::CliResult;
 use cargo::util::CargoResult;
 use cargo::util::Config;
+use metadata::CargoInternalsMetadataFetcher;
+use metadata::CargoWorkspaceFiles;
 use planning::BuildPlanner;
 use planning::BuildPlannerImpl;
-use planning::CargoInternalsMetadataFetcher;
-use planning::CargoWorkspaceFiles;
 use rendering::BuildRenderer;
 use rendering::FileOutputs;
 use rendering::RenderDetails;
@@ -138,11 +138,7 @@ fn real_main(options: Options, cargo_config: &Config) -> CliResult {
   };
 
   let dry_run = options.flag_dryrun.unwrap_or(false);
-  for FileOutputs {
-    path,
-    contents,
-  } in bazel_file_outputs
-  {
+  for FileOutputs { path, contents } in bazel_file_outputs {
     if !dry_run {
       try!(write_to_file_loudly(&path, &contents));
     } else {
